[issue #1720] Add support for AWS Organizations
ready for pull request
did a little cleanup refactoring
local tests pass
